{{cookiecutter.directory_name}}/pyproject.toml
[tool.poetry]
name = "dene"
version = "0.1.0"
description = ""
authors = ["kenan"]

[tool.poetry.dependencies]
python = "3.10.14"
numpy = ">=1.21.4,<2.0"
hydra-core = "^1.3.2"
opencv-python = "4.8.1.78"
pdoc3 = "^0.10.0"
pydicom = ">=2.4.3"
seaborn = "^0.13.0"
scikit-learn = "^1.3.1"
torch = { url = "https://download.pytorch.org/whl/cu121/torch-2.1.0%2Bcu121-cp310-cp310-linux_x86_64.whl" }
torchvision = { url = "https://download.pytorch.org/whl/cu121/torchvision-0.16.0%2Bcu121-cp310-cp310-linux_x86_64.whl" }
torchaudio = { url = "https://download.pytorch.org/whl/cu121/torchaudio-2.1.0%2Bcu121-cp310-cp310-linux_x86_64.whl" }
tensorboard = "2.15.1"
fastai = "^2.7.12"
iterative-stratification = "^0.1.7"
mlxtend = "^0.23.0"
monai = "^1.2.0"
jupyterlab = "^4.0.0"
torchsummary = "1.5.1"
mlflow = "^2.9.2"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"
dvc = "^3.42.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[tool.setuptools_scm]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
